refactor: common::Compat -> rpc::Compat
